+++ noether/catalogue/dimensions.py
@@ -1,3 +1,6 @@
+'''
+Various dimensions, for display purposes.
+'''
 
 from ..core.config import Config, conf
 from ..core.Dimension import dimensionless, Dimension
@@ -28,7 +31,7 @@ Config.register('dimension_angle', True, '''\
 Register angles as their own dimension, rather than dimensionless.''')
 
 if conf.get('dimension_angle'):
-    angle = Dimension.new("angle", "φ", 200)
+    angle = I(Dimension.new("angle", "φ", 200), "angle")
 else:
     angle = dimensionless
 solid_angle = I(angle**2, 'solid_angle')