+++ noether/catalogue/extension/__init__.py
@@ -1,0 +1,8 @@
+from .historic import *
+from .cgs import *
+from .typographic import *
+
+from .regional import *
+from .ancient.biblical import *
+
+# TODO: Many, many, many others