+++ CHANGELOG.md
@@ -6,6 +6,10 @@
     - You can chain units with `&`. By default (`conf.human_time`), for example, time is displayed in `year & day & hour & minute & second`.
     - Support for `rich` API (automatic in `python -im noether`)
 
+- Expanded catalogue
+    - More obscure, regional and historic units added
+    - `catalogue_extended` can be disabled to bypass historical units for speedup
+
 - Core code rewritten entirely from scratch
     - `Measure` is now frozen, and is generic to its underlying value, which can be any `Real` subclass. `Measure.cast(T)` can be used to cast to a type.
     - `Measure.stddev` is `None` by default rather than zero.